Added niemyjski's fix for callbacks on close
Callback timer is now stopped when the dialog is closed. Normally, this
happens anyway when the object is disposed when the dialog gets
disposed. However, the dialog raises an event that consumers can handle.
During this time, the timer was still running since it hadn't gotten
cleaned up yet. It is now being explicitly stopped.