Common buttons localisation; Cancel-only button
Added static properties to the TaskDialog class that allow providing localised texts for the common buttons like Cancel or Retry. Set them once before using a TaskDialog. If null, the English default will be used as before.

Also, using only the Cancel button of the common buttons is now possible.